Added ELF symbol loading and refactored loader
Added a symbol loading interface which allows for loading of other
symbols than a stack. An event and a main stack is now being loaded.
These have also been added to the linker script.

The ELF loader code was refactored to match the style of the rest of the
codebase.

Some more error reporting was added in CEmulatorHost::CreateFromProgram
function. The CEmulatorHost constructor was also refactored to no
longer take an Executable struct refrence as a parameter.

Pulled changes from MIPS emulator