Implement Press and Release Gestures
Data is sent to c++ through the property grid_gesture_data, whose
changes is listened to in Grid.h HandleDiffs.

It uses the rather simple protocol:

identifier/action type (1 byte)

index (1 byte)

the identifier is currently only for press (0x01) and release (0x00).
index is the grid tile index that was interacted with. it is simply a
flattened 2d array.

These two values are then encoded as hex, with a salt added to ensure
a property change is guaranteed to be recognized by the host.

Two tests are included, grid_test.c and grid_paint.c.